requirements.txt
Flask==2.0.3  # Use a version of Flask that was compatible with Werkzeug 2.0.3
requests==2.28.1
python-dotenv==0.21.0
redis==4.5.4
requests_oauthlib==1.3.1
werkzeug==2.0.3  # Downgrade werkzeug
pydantic==1.10.2
SQLAlchemy==1.4.46  # Compatible version
flask-sqlalchemy==2.5.1
flask-migrate==3.1.0
psycopg2-binary==2.9.6
